requirements.txt
diffusers>=0.30.0
transformers>=4.44.0
accelerate
torch
torchvision
gradio
numpy
pillow
dashscope
spaces
python-dotenv
